requirements.txt
bencoder.pyx>=1.1.2
humanfriendly
tqdm>=4.8.4